+++ app/Http/Controllers/UserController.php
@@ -3,22 +3,29 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UserFormRequest;
+use App\Models\Reserve;
 use App\Models\User;
 use Exception;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
     public function index()
     {
-        $users = User::all();
+        try {
+            $users = User::all();
 
-        $messageSuccess = session('messageSuccess');
+            $messageSuccess = session('messageSuccess');
 
-        return view('users.index')
-            ->with('users', $users)
-            ->with('messageSuccess', $messageSuccess);
+            return view('users.index')
+                ->with('users', $users)
+                ->with('messageSuccess', $messageSuccess);
+
+        } catch(Exception $e) {
+            return back()->withErrors($e->getMessage());
+        }
     }
 
     public function create()
@@ -69,6 +76,10 @@ class UserController extends Controller
     public function destroy(Request $request)
     {
         try {
+            if (Reserve::where('users_id', $request->user)) {
+                throw new Exception('Não é possível remover usuários que possuem reservas cadastradas');
+            } 
+
             User::destroy($request->user);
 
             $request->session()->flash('messageSuccess', 'Usuário removido com sucesso');
